chore: rewrote ShiftCRUD to a func component
